refactor: muting copies to common file
